+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'auth/login.dart';
+import 'services/session_storage.dart';
+import 'dashboard.dart';
 
 void main() {
   runApp(const MainApp());
@@ -10,14 +12,20 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
+    return MaterialApp(
       home: Scaffold(
         body: Center(
-          child: LogInWidget(),
+          child: FutureBuilder<bool>(
+            future: checkSession(),
+            builder: (context, snapshot) {
+              if (snapshot.connectionState == ConnectionState.waiting) {
+                return const CircularProgressIndicator();
+              }
+              return snapshot.data ?? true ? const Dashboard() : const LogInWidget();
+            },
+          ),
         ),
       ),
     );
   }
 }
-
-//TODO - Rafa: Agregar los cambios de pantallas entre widgets