+++ lib/user-config.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class userConfig extends StatelessWidget {
+  const userConfig({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: Center(
+          child: Text("Hola, mundo"),
+        ),
+      ),
+    );
+  }
+}